+++ simple-stack-example-multistack/src/main/java/com/zhuinden/simplestackdemomultistack/application/Key.java
@@ -4,13 +4,19 @@ import android.content.Context;
 import android.os.Parcelable;
 
 import com.zhuinden.simplestack.BackstackDelegate;
+import com.zhuinden.simplestackdemomultistack.util.ServiceLocator;
 
 /**
  * Created by Owner on 2017. 01. 12..
  */
 
-public interface Key extends Parcelable {
-    int layout();
+public abstract class Key
+        implements Parcelable {
+    public abstract int layout();
 
-    BackstackDelegate selectDelegate(Context context);
+    public final BackstackDelegate selectDelegate(Context context) {
+        return ServiceLocator.getService(context, stackIdentifier());
+    }
+
+    public abstract String stackIdentifier();
 }