+++ simple-stack-example-multistack/src/main/java/com/zhuinden/simplestackdemomultistack/presentation/paths/main/chromecast/ChromeCastKey.java
@@ -1,13 +1,9 @@
 package com.zhuinden.simplestackdemomultistack.presentation.paths.main.chromecast;
 
-import android.content.Context;
-
 import com.google.auto.value.AutoValue;
-import com.zhuinden.simplestack.BackstackDelegate;
 import com.zhuinden.simplestackdemomultistack.R;
 import com.zhuinden.simplestackdemomultistack.application.Key;
 import com.zhuinden.simplestackdemomultistack.application.MainActivity;
-import com.zhuinden.simplestackdemomultistack.util.ServiceLocator;
 
 /**
  * Created by Owner on 2017. 01. 12..
@@ -15,7 +11,7 @@ import com.zhuinden.simplestackdemomultistack.util.ServiceLocator;
 
 @AutoValue
 public abstract class ChromeCastKey
-        implements Key {
+        extends Key {
     @Override
     public int layout() {
         return R.layout.path_chromecast;
@@ -26,7 +22,7 @@ public abstract class ChromeCastKey
     }
 
     @Override
-    public BackstackDelegate selectDelegate(Context context) {
-        return ServiceLocator.getService(context, MainActivity.StackType.CHROMECAST.name());
+    public String stackIdentifier() {
+        return MainActivity.StackType.CHROMECAST.name();
     }
 }