+++ simple-stack-example-multistack/src/main/java/com/zhuinden/simplestackdemomultistack/presentation/paths/main/cloudsync/another/AnotherKey.java
@@ -1,4 +1,4 @@
-package com.zhuinden.simplestackdemomultistack.presentation.paths.main.cloudsync;
+package com.zhuinden.simplestackdemomultistack.presentation.paths.main.cloudsync.another;
 
 import com.google.auto.value.AutoValue;
 import com.zhuinden.simplestackdemomultistack.R;
@@ -6,22 +6,22 @@ import com.zhuinden.simplestackdemomultistack.application.Key;
 import com.zhuinden.simplestackdemomultistack.application.MainActivity;
 
 /**
- * Created by Owner on 2017. 01. 12..
+ * Created by Zhuinden on 2017.02.19..
  */
 @AutoValue
-public abstract class CloudSyncKey
+public abstract class AnotherKey
         extends Key {
     @Override
     public int layout() {
-        return R.layout.path_cloudsync;
-    }
-
-    public static CloudSyncKey create() {
-        return new AutoValue_CloudSyncKey();
+        return R.layout.path_another;
     }
 
     @Override
     public String stackIdentifier() {
         return MainActivity.StackType.CLOUDSYNC.name();
+    }
+
+    public static Object create() {
+        return new AutoValue_AnotherKey();
     }
 }