+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change log
 
+-Simple Stack 1.1.1 (2017-02-19)
+--------------------------------
+- FIX: A bug that allowed the possibility that an uninitialized backstack could restore its history to be an empty list after process death.
+
+- ADDED: `simple-stack-example-multistack` for handling multiple backstacks in the same Activity using a BottomNavigationView.
+
 -Simple Stack 1.1.0 (2017-02-18)
 --------------------------------
 - BREAKING CHANGE: `Backstack`'s APIs return `Object` instead of `Parcelable` (that includes `StateChange`, initial keys, `HistoryBuilder`, etc.)